pyproject.toml
[project]
name = "topicos_avancados_2026_1_equipe_jud_3_atividade2"
version = "0.1.0"
description = "Projeto de avaliação de LLMs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "datasets>=4.8.5",
    "openai>=1.50.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scipy>=1.14.1",
    "typer>=0.25.1",
    "yoyo-migrations>=9.0.0",
]

[project.scripts]
app = "src.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "complexipy>=5.2.0",
    "py2puml>=0.11.0",
    "ruff>=0.15.7",
]

[tool.complexipy]
paths = ["src", "tests"]
max-complexity-allowed = 10
snapshot-create = false
snapshot-ignore = false
quiet = false
ignore-complexity = false
failed = false
color = "auto"
sort = "asc"
exclude = []
output-csv = false
output-json = false
